Refactor BLE_PURGE_MIN_AGE from 8 bit to 16

+++ main/bluetooth.h
@@ -66,7 +66,7 @@ extern uint8_t gravity_sel_bt_count;
 
 extern const char *BT_TAG;
 extern gravity_bt_purge_strategy_t purgeStrategy;
-extern uint8_t PURGE_MIN_AGE;
+extern uint16_t PURGE_MIN_AGE;
 extern int32_t PURGE_MAX_RSSI;
 
 esp_err_t gravity_ble_scan_start(gravity_bt_purge_strategy_t purgeStrat);
